Snake movements
Basically just pop off the tail and put it as the head in the new position. For now it happens at a super crazy speed because requestAnimationFrame is doing it's little optimisation thing. We'll have to get the time difference and use that to only update after a certain period, or some other time based solution.