Add game over condition
The game is over if the snake goes outside the board on any of the four ends of if it touches itself. Added a game over flag so the game could stop running and to render game over to the user. Made the column and row variables properties of the Snake game so I can test the boundaries w/o using pixels.